+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Update Beast subtree to v41
 * Update Travis targets and build scripts
+* Update example programs
 
 --------------------------------------------------------------------------------